Updates har.js to fix dates and adjust totalTime
Updated the calculation of the dates based on the work from Benedict.

I updated the calculation of totalTime, and refactored the calculation of receiveTime.
The previous totalTime ignored blockedTime, which it should not have. I can't just add blockedTime to the old expression because in the case of the first resource, the blockedTime is equal to timing.sendStart (because connectStart is invariably 0) given there's no DNS lookup.