Fixed the per resource startedDateTime
The startedDateTime should simply be the object.requestMessage.wallTime*1000.
The object.requestMessage.timestamp should not have been subtracted to it as was done in 8cac3fd, as this timestamp is generated with javascript monotonically increasing timestamps, while wallTime is absolute(ish).
The object.requestMessage.timestamp can only be used with a reference timestamp generated by the same function that generated it.

After implementing this, startedDateTime per resource now makes more sense. With timing.requestTime, it provided dates in the 1970, with the implementation in 8cac3fd, it jumped 10s of minutes, which was not true in the test being conducted.